refactor: consolidate TypeArea type and update component imports
