+++ src/types/query-overpass.d.ts
@@ -1,0 +1,4 @@
+declare module '@derhuerst/query-overpass' {
+  function queryOverpass(query: string): Promise<any[]>;
+  export default queryOverpass;
+}
